road generation now remembers how the road ended in the previous chunk so that it can flow better

+++ src/main/java/jMonkeyEngine/Road/Node.java
@@ -11,7 +11,7 @@ public class Node implements Comparable<Node>, Serializable {
     public float gCost; // cost from start
     public float fCost; // gCost + heuristic
     public Node parent;
-    public int dxFromParent, dyFromParent;
+    public float dxFromParent, dyFromParent;
     public float dirMag;
 
     public Node(int x, int y) {
